builder:siwisdk: check existance of main.cpp
Check both main.c and main.cpp before creating template

+++ builder/framework/siwisdk.py
@@ -19,7 +19,8 @@ assert isdir(FRAMEWORK_DIR)
 
 # Create Project Template
 main_c = join(env.subst("$PROJECT_DIR"), "src", "main.c")
-if False == isfile(main_c):
+main_cpp = join(env.subst("$PROJECT_DIR"), "src", "main.cpp")
+if (False == isfile(main_c)) and (False == isfile(main_cpp)):
     copyfile(join(FRAMEWORK_DIR, "template", "main.c"), main_c)
 
 def fota_crc16(data:bytearray, length):
